Crate problem055

Problem 55 (Lychrel numbers)

Solution detail

Simply check each number up to 10,000 to see if it is a Lychrel number, remembering to use arbitrary-precision integers since the numbers involved can grow very large.

Functions

answer
Solve the problem, returning the answer as a String

is_lychrel
Determines whether the given number is a possible Lychrel number, as determined by performing the reverse-and-add iteration up to the given threshold number of iterations.

main

reverse
Calculates the result of reversing the digits of the given number.

solve
Find the number of Lychrel numbers below the given limit, as determined by performing the reverse-and-add iteration up to the given threshold number of iterations.